ctl-genmanifest has been rewritten as ctl-build-component.
ctl-genmanifest was written with optparse, was hard to maintain and had a
problem with component documentation files. Due to that it was rewritten as
ctl-build-component which has a new Feature: ctl-build-component is independet
from the ctlweb backend libraries. It only depends for the sake of easy
deployment on the standard python3.2 libraries.